requirements.txt
protobuf
streamlit
numpy
scikit-learn
python-dotenv
pandas
google-generativeai